requirements.txt
attrs==19.3.0
markdown2==2.3.8
python-dotenv==0.10.3
pyyaml==5.2
requests==2.22.0
tqdm==4.41.0
